♻️ Separate tests and test cases (#845)
